requirements.txt
Django==4.0.4
Pillow~=9.1.0